Added a check on attribute availability when retrieving index form text line to prevent index out of range.
Fix is mainly to cater for pythonanywhere's old tesseract version where some of the attributes can not be recognized. Also to make the codes more robust. The extraction of attributes index value from the text line previously was not checked, which resulted in the index out of range error when the attributes are not converted properly by tesseract.